// src/components/Modal/DetailModal/utils/type.ts
export interface detailDataType {
  addressName: string;
  categoryGroupCode: string;
  categoryGroupName: string;
  distance: string;
  phone: string;
  placeId: string;
  placeName: string;
  placeUrl: string;
  reviews: number;
  roadAddressName: string;
  views: number;
  x: string;
  y: string;
}